+++ src/main/java/com/selenium/pp/pages/BasePage.java
@@ -1,4 +1,4 @@
-package com.page;
+package com.selenium.pp.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
