+++ src/test/java/com/page/pomPP/AddressBookPageTest.java
@@ -1,11 +1,13 @@
-package com.page;
+package com.page.pomPP;
 
-import models.Address;
-import models.User;
+import com.unknown.models.Address;
+import com.unknown.models.User;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.testng.annotations.Test;
 import runner.BaseTest;
+import com.selenium.pp.pages.AddressBookPage;
+import com.selenium.pp.pages.MainPage;
 
 public class AddressBookPageTest extends BaseTest {
