+++ src/test/java/selpw/FirstTest.java
@@ -1,3 +1,5 @@
+package selpw;
+
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import runner.BaseTest;
