+++ src/test/java/selpw/LumaIkTest.java
@@ -1,3 +1,5 @@
+package selpw;
+
 import com.microsoft.playwright.Locator;
 import com.microsoft.playwright.Page;
 import org.openqa.selenium.By;
