+++ src/test/java/selpw/SearchBoxTest.java
@@ -1,3 +1,5 @@
+package selpw;
+
 import org.openqa.selenium.By;
 import org.testng.annotations.Test;
 import runner.BaseTest;
